Issue 496 Backend env route availability
