+++ backend/sbom.json
@@ -6728,7 +6728,7 @@
       ],
       "name": "nachet-backend",
       "type": "application",
-      "version": "2.3.0"
+      "version": "2.4.0"
     },
     "properties": [
       {
